+++ lib/nano_rpc.rb
@@ -7,16 +7,17 @@ require 'nano_rpc/helpers/accounts_helper'
 require 'nano_rpc/helpers/node_helper'
 require 'nano_rpc/helpers/wallet_helper'
 
-require 'nano_rpc/proxy_methods/account'
-require 'nano_rpc/proxy_methods/accounts'
-require 'nano_rpc/proxy_methods/node'
-require 'nano_rpc/proxy_methods/wallet'
+require 'nano_rpc/methods/account_methods'
+require 'nano_rpc/methods/accounts_methods'
+require 'nano_rpc/methods/node_methods'
+require 'nano_rpc/methods/wallet_methods'
 
 require 'nano_rpc/proxy'
+require 'nano_rpc/proxies/account'
+require 'nano_rpc/proxies/accounts'
+require 'nano_rpc/proxies/wallet'
+
 require 'nano_rpc/node'
 require 'nano_rpc/errors'
 require 'nano_rpc/response'
 require 'nano_rpc/numeric'
-require 'nano_rpc/proxy_objects/account'
-require 'nano_rpc/proxy_objects/accounts'
-require 'nano_rpc/proxy_objects/wallet'
