+++ lib/nano_rpc/methods/account_methods.rb
