+++ lib/nano_rpc/methods/accounts_methods.rb
