+++ lib/nano_rpc/methods/node_methods.rb
