+++ lib/nano_rpc/methods/wallet_methods.rb
