+++ lib/nano_rpc/proxies/account.rb
