+++ lib/nano_rpc/proxies/wallet.rb
